backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-api-python-client==2.131.0
google-auth==2.29.0
python-dateutil==2.9.0
pytz==2024.1
pydantic>=2.7.0
requests>=2.31.0
pytest>=8.0.0
httpx>=0.27.0
